#3 Change to check the email
I found there are no way to get the email of seller with current exist method. So the require method return the email of seller from sellerDoa, or better yet, return a seller object form sellerDoa with sellerID as argument. Is there going check the email is repeated during register? Btw, the code can not pass the compile on my pc even before the change in this commit.