#4 Implementation of email checking
Adding method getCustomerID() and getSellerID() in seller and customer DOA class. This two method receive a String of email and return the ID, if not found return -1. Using this method, we can check an email being used or not by invoke this two method. If -1 is returned, mentioned the user and try to input another one